tests/wpt/css-tests/css-text-3_dev/html/css3-text-line-break-baspglwj-120.htm

<!DOCTYPE html>
<html lang="en"><head>
<meta charset="utf-8">
<title>CSS3 Text, linebreaks: U+00A0 NO-BREAK SPACE</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida">
<link href="https://drafts.csswg.org/css-text-3/#line-break-details" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="font dom" name="flags">
<meta content="[Exploratory] The browser will NOT break a line of text containing any Unicode character with the GL non-breaking property." name="assert">
<style type="text/css">
@font-face {
    font-family: 'csstest_ascii';
    src: url('support/csstest-ascii-webfont.woff') format('woff');
    font-weight: normal;
    font-style: normal;
	}
#breakable { font-family: csstest_ascii; font-size: 25px; width: 800px; line-height: 30px; }
</style>
</head>
<body>



<div class="test">
			<div id="breakable">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&nbsp;bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
			</div>


<!--Notes:

Box height: <span id='debugresult'></span><script>document.getElementById('debugresult').innerHTML = document.getElementById('breakable').offsetHeight;</script>

-->
<script>
test(function() {
assert_true(document.getElementById('breakable').offsetHeight < 35);
}, " ");
</script>

<div id="log"></div>



</body></html>
